* Major overhaul of linux build system ** Only openssl and bz2 (for stuffextract) left as external dependencies, zlib, zthread and irrlicht are compiled from dep folder ** Irrlicht upgraded to version 1.5 ** Various fixes to make GCC not complain ** to compile: autoreconf -if && ./compile -prefix=full_path_to_bin_directory && make ** Only issue: rename dep/openssl otherwise compiler will complain - too late to figure out now.
git-svn-id: https://pseuwow.svn.sourceforge.net/svnroot/pseuwow@351 fda9a0ab-4e26-0410-ba80-894f06a4b443

+++ src/dep/src/irrlicht/CTRTextureGouraudNoZ.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2002-2007 Nikolaus Gebhardt
+// Copyright (C) 2002-2009 Nikolaus Gebhardt
 // This file is part of the "Irrlicht Engine".
 // For conditions of distribution and use, see copyright notice in irrlicht.h
 
@@ -103,8 +103,7 @@ public:
 			if (!TriangleRect.isRectCollided(ViewPortRect))
 				continue;
 
-
-			// höhe des dreiecks berechnen
+			// calculate height of triangle
 			height = v3->Pos.Y - v1->Pos.Y;
 			if (!height)
 				continue;
@@ -130,17 +129,17 @@ public:
 			{
 				tmpDiv = 1.0f / (f32)(v2->Pos.Y - v1->Pos.Y);
 				rightdeltaxf = (v2->Pos.X - v1->Pos.X) * tmpDiv;
-				rightStepR = (s32)(((video::getRed(v2->Color)<<8) - rightR) * tmpDiv);
-				rightStepG = (s32)(((video::getGreen(v2->Color)<<8) - rightG) * tmpDiv);
-				rightStepB = (s32)(((video::getBlue(v2->Color)<<8) - rightB) * tmpDiv);
+				rightStepR = (s32)(((s32)(video::getRed(v2->Color)<<8) - rightR) * tmpDiv);
+				rightStepG = (s32)(((s32)(video::getGreen(v2->Color)<<8) - rightG) * tmpDiv);
+				rightStepB = (s32)(((s32)(video::getBlue(v2->Color)<<8) - rightB) * tmpDiv);
 				rightTxStep = (s32)((v2->TCoords.X - rightTx) * tmpDiv);
 				rightTyStep = (s32)((v2->TCoords.Y - rightTy) * tmpDiv);
 
 				tmpDiv = 1.0f / (f32)height;
 				leftdeltaxf = (v3->Pos.X - v1->Pos.X) * tmpDiv;
-				leftStepR = (s32)(((video::getRed(v3->Color)<<8) - leftR) * tmpDiv);
-				leftStepG = (s32)(((video::getGreen(v3->Color)<<8) - leftG) * tmpDiv);
-				leftStepB = (s32)(((video::getBlue(v3->Color)<<8) - leftB) * tmpDiv);
+				leftStepR = (s32)(((s32)(video::getRed(v3->Color)<<8) - leftR) * tmpDiv);
+				leftStepG = (s32)(((s32)(video::getGreen(v3->Color)<<8) - leftG) * tmpDiv);
+				leftStepB = (s32)(((s32)(video::getBlue(v3->Color)<<8) - leftB) * tmpDiv);
 				leftTxStep = (s32)((v3->TCoords.X - leftTx) * tmpDiv);
 				leftTyStep = (s32)((v3->TCoords.Y - leftTy) * tmpDiv);
 			}
@@ -148,17 +147,17 @@ public:
 			{
 				tmpDiv = 1.0f / (f32)height;
 				rightdeltaxf = (v3->Pos.X - v1->Pos.X) * tmpDiv;
-				rightStepR = (s32)(((video::getRed(v3->Color)<<8) - rightR) * tmpDiv);
-				rightStepG = (s32)(((video::getGreen(v3->Color)<<8) - rightG) * tmpDiv);
-				rightStepB = (s32)(((video::getBlue(v3->Color)<<8) - rightB) * tmpDiv);
+				rightStepR = (s32)(((s32)(video::getRed(v3->Color)<<8) - rightR) * tmpDiv);
+				rightStepG = (s32)(((s32)(video::getGreen(v3->Color)<<8) - rightG) * tmpDiv);
+				rightStepB = (s32)(((s32)(video::getBlue(v3->Color)<<8) - rightB) * tmpDiv);
 				rightTxStep = (s32)((v3->TCoords.X - rightTx) * tmpDiv);
 				rightTyStep = (s32)((v3->TCoords.Y - rightTy) * tmpDiv);
 
 				tmpDiv = 1.0f / (f32)(v2->Pos.Y - v1->Pos.Y);
 				leftdeltaxf = (v2->Pos.X - v1->Pos.X) * tmpDiv;
-				leftStepR = (s32)(((video::getRed(v2->Color)<<8) - leftR) * tmpDiv);
-				leftStepG = (s32)(((video::getGreen(v2->Color)<<8) - leftG) * tmpDiv);
-				leftStepB = (s32)(((video::getBlue(v2->Color)<<8) - leftB) * tmpDiv);
+				leftStepR = (s32)(((s32)(video::getRed(v2->Color)<<8) - leftR) * tmpDiv);
+				leftStepG = (s32)(((s32)(video::getGreen(v2->Color)<<8) - leftG) * tmpDiv);
+				leftStepB = (s32)(((s32)(video::getBlue(v2->Color)<<8) - leftB) * tmpDiv);
 				leftTxStep = (s32)((v2->TCoords.X - leftTx) * tmpDiv);
 				leftTyStep = (s32)((v2->TCoords.Y - leftTy) * tmpDiv);
 			}
@@ -252,7 +251,9 @@ public:
 						while (hSpanBegin < hSpanEnd) 
 						{
 							color = lockedTexture[((spanTy>>8)&textureYMask) * lockedTextureWidth + ((spanTx>>8)&textureXMask)];
-							*hSpanBegin = video::RGB16(video::getRed(color) * (spanR>>8) >>2, video::getGreen(color) * (spanG>>8) >>2, video::getBlue(color) * (spanB>>8) >>2);
+							*hSpanBegin = video::RGB16(video::getRed(color) * (spanR>>8) >>2,
+									video::getGreen(color) * (spanG>>8) >>2,
+									video::getBlue(color) * (spanB>>8) >>2);
 
 							spanR += spanStepR;
 							spanG += spanStepG;
@@ -299,9 +300,9 @@ public:
 					rightR = video::getRed(v2->Color)<<8;
 					rightG = video::getGreen(v2->Color)<<8;
 					rightB = video::getBlue(v2->Color)<<8;
-					rightStepR = (s32)(((video::getRed(v3->Color)<<8) - rightR) * tmpDiv);
-					rightStepG = (s32)(((video::getGreen(v3->Color)<<8) - rightG) * tmpDiv);
-					rightStepB = (s32)(((video::getBlue(v3->Color)<<8) - rightB) * tmpDiv);
+					rightStepR = (s32)(((s32)(video::getRed(v3->Color)<<8) - rightR) * tmpDiv);
+					rightStepG = (s32)(((s32)(video::getGreen(v3->Color)<<8) - rightG) * tmpDiv);
+					rightStepB = (s32)(((s32)(video::getBlue(v3->Color)<<8) - rightB) * tmpDiv);
 
 					rightTx = v2->TCoords.X;
 					rightTy = v2->TCoords.Y;
@@ -318,9 +319,9 @@ public:
 					leftR = video::getRed(v2->Color)<<8;
 					leftG = video::getGreen(v2->Color)<<8;
 					leftB = video::getBlue(v2->Color)<<8;
-					leftStepR = (s32)(((video::getRed(v3->Color)<<8) - leftR) * tmpDiv);
-					leftStepG = (s32)(((video::getGreen(v3->Color)<<8) - leftG) * tmpDiv);
-					leftStepB = (s32)(((video::getBlue(v3->Color)<<8) - leftB) * tmpDiv);
+					leftStepR = (s32)(((s32)(video::getRed(v3->Color)<<8) - leftR) * tmpDiv);
+					leftStepG = (s32)(((s32)(video::getGreen(v3->Color)<<8) - leftG) * tmpDiv);
+					leftStepB = (s32)(((s32)(video::getBlue(v3->Color)<<8) - leftB) * tmpDiv);
 
 					leftTx = v2->TCoords.X;
 					leftTy = v2->TCoords.Y;
